refactor(core): извлечь ядро sessions (create/complete/review)
